requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
Django==1.11.29
django-filter==2.0.0
djangorestframework==3.9.4
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pytz==2020.5
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
